+++ SpaceBattle.Lib/ICommand.cs
@@ -1,0 +1,6 @@
+﻿namespace SpaceBattle.Lib;
+public interface ICommand
+{
+    public void Execute();
+
+}
