+++ SpaceBattle.Lib/IMoving.cs
@@ -1,0 +1,7 @@
+﻿namespace SpaceBattle.Lib;
+
+public interface IMoving
+{
+    Vec Position { get; set; }
+    Vec Velocity { get; }
+}
